refactor(navbar): replace navbarhome with common navbar component
