+++ src/components/users/home/Home.jsx
@@ -1,16 +1,16 @@
+import Navbar from "../../common/Navbar";
 import Article from "./Article";
 import Carousel from "./Carousel";
 import CartButton from "./CartButton";
 import Category from "./Category";
 import CheapersProduct from "./CheapestProduct";
-import NavbarHome from "./NavbarHome";
 import PopularProduct from "./PopularProduct";
 import PosterCard from "./PosterCard";
 
 export default function Home() {
   return (
     <>
-      <NavbarHome />
+      <Navbar />
       <Carousel />
       <Category />
       <PosterCard />
